Refactor: 레이어 아키텍처 구성
