+++ src/main/java/io/hhplus/tdd/ApiControllerAdvice.java
@@ -1,6 +1,5 @@
 package io.hhplus.tdd;
 
-import io.hhplus.tdd.point.BizException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
