+++ src/main/java/io/hhplus/tdd/domain/point/PointHistory.java
@@ -1,4 +1,6 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.domain.point;
+
+import io.hhplus.tdd.interfaces.api.point.TransactionType;
 
 public record PointHistory(
         long id,
