+++ src/main/java/io/hhplus/tdd/interfaces/api/point/PointRequest.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.interfaces.api.point;
 
 public record PointRequest(
         long amount
