+++ src/main/java/io/hhplus/tdd/interfaces/api/point/TransactionType.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.interfaces.api.point;
 
 
 /**
